requirements.txt
Django
djangorestframework
djangorestframework-jwt
Pillow
django-rest-swagger